requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.0.11
colorama==0.4.4
colorlog==6.6.0
discord-py-interactions==3.0.2
discord-py-slash-command==3.0.3
discord.py @ git+https://github.com/Rapptz/discord.py@45d498c1b76deaf3b394d17ccf56112fa691d160
frozenlist==1.3.0
hikari==2.0.0.dev105
idna==3.3
multidict==5.2.0
pure25519==0.0.1
python-dotenv==0.19.2
typing-extensions==4.0.1
yarl==1.7.2